@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             v4.25.1
 // source: hello.proto
 
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	GreeterService_SayHello_FullMethodName = "/App.Grpc.Hello.GreeterService/SayHello"
@@ -25,6 +25,8 @@ const (
 // GreeterServiceClient is the client API for GreeterService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// service 定义开放调用的服务
 type GreeterServiceClient interface {
 	SayHello(ctx context.Context, in *HelloReq, opts ...grpc.CallOption) (*HelloReply, error)
 }
@@ -38,8 +40,9 @@ func NewGreeterServiceClient(cc grpc.ClientConnInterface) GreeterServiceClient {
 }
 
 func (c *greeterServiceClient) SayHello(ctx context.Context, in *HelloReq, opts ...grpc.CallOption) (*HelloReply, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(HelloReply)
-	err := c.cc.Invoke(ctx, GreeterService_SayHello_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, GreeterService_SayHello_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -48,20 +51,26 @@ func (c *greeterServiceClient) SayHello(ctx context.Context, in *HelloReq, opts 
 
 // GreeterServiceServer is the server API for GreeterService service.
 // All implementations must embed UnimplementedGreeterServiceServer
-// for forward compatibility
+// for forward compatibility.
+//
+// service 定义开放调用的服务
 type GreeterServiceServer interface {
 	SayHello(context.Context, *HelloReq) (*HelloReply, error)
 	mustEmbedUnimplementedGreeterServiceServer()
 }
 
-// UnimplementedGreeterServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedGreeterServiceServer struct {
-}
+// UnimplementedGreeterServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedGreeterServiceServer struct{}
 
 func (UnimplementedGreeterServiceServer) SayHello(context.Context, *HelloReq) (*HelloReply, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method SayHello not implemented")
 }
 func (UnimplementedGreeterServiceServer) mustEmbedUnimplementedGreeterServiceServer() {}
+func (UnimplementedGreeterServiceServer) testEmbeddedByValue()                        {}
 
 // UnsafeGreeterServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to GreeterServiceServer will
@@ -71,6 +80,13 @@ type UnsafeGreeterServiceServer interface {
 }
 
 func RegisterGreeterServiceServer(s grpc.ServiceRegistrar, srv GreeterServiceServer) {
+	// If the following call pancis, it indicates UnimplementedGreeterServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&GreeterService_ServiceDesc, srv)
 }
 
